requirements.txt
annotated-types==0.6.0
app-store-scraper==0.3.5
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gspread==6.1.0
httplib2==0.22.0
idna==2.10
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.8.4
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.5.15
requests==2.23.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.25.11
vaderSentiment==3.3.2
wordcloud==1.9.3
zipp==3.18.1
